pages を admin に Renama し EditorComponent を移動
