+++ src/app/admin/components/editor/editor.component.html
