+++ src/app/admin/components/editor/editor.component.scss
@@ -1,4 +1,4 @@
-@import '../../shared/styles/common';
+@import '../../../shared/styles/common';
 
 .editor {
   display: flex;
