+++ src/app/admin/components/editor/editor.component.spec.ts
